switch to using erb for the home page
Rendering simply text or html from the Sinatra `app.rb` file is certainly not the best way to go about serving HTML documents in Sinatra.

It's preferred that we use a templating language to display HTML. The templating language that comes with Sinatra is called ERB. With `ERB`, we're able to write HTML as in standard `.html` files in addition to putting Ruby code in there as well. So we're basically able to add Ruby login to otherwise logicless HTML.

To user ERB we must create a folder called `views` that is in the directory as the main Sinatra app files which is `app.rb` in our case.

In order to render the ERB template we simply add the following line to our action:
```ruby
erb :index
```
Not here that `erb` is a method that comes from the Sinatra gem, so this can be written as:
```ruby
erb(:index)
```