build the temperature conversion action
We need to capture the parameters to convert the temperature from Celsius to Fahrenheit.

We make the actions by in our `app.rb` with `post "/convert"` because the form we build in the previous step has a method of `post` and an action of `/convert`.

We get the parameters submitted by user in a `params` object available in any action within our `app.rb` file. You can think of the `params` as a ruby Hash whose keys are the `name` attributes given in the form and values are the entries provided by the user.

In our case, we have the `name` attribute for our input as `temparature` so if the user enters 15 in the temperature field the params hash will look like:
```ruby
{"name" => "15"}
```
Couple of things to note:
* The values will always be strings so we will have to do a conversion if have other data types such as floats or strings, this is why we have the line to convert the parameter into a float:
```ruby
temperature_in_c = params[:temperature].to_f
```
* We're able to access the values of the Hash with either symbols or strings so both `params[:temperature]` and `params["temperature"]` work.

We then calculate the Fahrenheit temperature and then we just render it as a string for now.