make the result show on the same page
We improve our solution now by making the result show on the same page. So instead of simply rendering the result as a text we just re-render the index ERB template: `erb :index`.

In order to access the values in the ERB template we will need to make them instance variables so we switch the `temperature_in_c` to `@temperature_in_c` and the `temperature_in_f` to `@temperature_in_f`.

We then use the ERB blocks to dispaly the results in Fahrenheit if it exists. The reason why we need to do so is because we use the same file `index.erb` for both actions: `get "/"` and `post "/convert"` and the `@temperature_in_f` doesn't exist in the first action.

We also display the value in the input field by simply setting the `value` attribute on the HTML tag. For the `get "/"` action, the `@temperature_in_f` will be `nil` so it will just show nothing.